requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.0.344
langchain-openai==0.0.2
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
python-multipart==0.0.6
pypdf==3.17.1